Eliminate the goto_next_pick command
The loom does not support unsolicited "go to next pick" (=C) commands.
This means jump_to_pick has to be queued up until the loom
requests the next pick. And that adds a lot of complexity.

Add a button to the debug area to send the "next pick" oob command.